// examples/effect/src/gen/effect/Order.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import * as DateTime from 'effect/DateTime'
import * as Schema from 'effect/Schema'

export type Order = {
  readonly id?: bigint
  readonly petId?: bigint
  readonly quantity?: number
  readonly shipDate?: DateTime.Utc
  readonly status?: 'placed' | 'approved' | 'delivered'
  readonly http_status?: 200 | 400 | 500
  readonly value?: 0 | 1 | 2 | 3 | 3.5 | 4
  readonly complete?: boolean
}

export const Order = Schema.Struct({
  id: Schema.optionalKey(Schema.BigInt.annotate({ format: 'int64', examples: [BigInt('10')] })),
  petId: Schema.optionalKey(Schema.BigInt.annotate({ format: 'int64', examples: [BigInt('198772')] })),
  quantity: Schema.optionalKey(Schema.Number.check(Schema.isFinite(), Schema.isInt()).annotate({ format: 'int32', examples: [7] })),
  shipDate: Schema.optionalKey(Schema.DateTimeUtcFromString.annotate({ format: 'date-time' })),
  status: Schema.optionalKey(Schema.Literals(['placed', 'approved', 'delivered']).annotate({ description: 'Order Status', examples: ['approved'] })),
  http_status: Schema.optionalKey(Schema.Literals([200, 400, 500]).annotate({ description: 'HTTP Status', examples: [200] })),
  value: Schema.optionalKey(Schema.Literals([0, 1, 2, 3, 3.5, 4]).annotate({ description: 'Price', examples: [2] })),
  complete: Schema.optionalKey(Schema.Boolean),
})