// examples/effect/src/gen/effect/Pet.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import * as Schema from 'effect/Schema'
import { Category } from './Category'
import { Tag } from './Tag'

export type Pet = {
  readonly id?: bigint
  readonly parent?: ReadonlyArray<Pet>
  readonly internalId?: string
  readonly name: string
  readonly category?: Category
  readonly photoUrls: ReadonlyArray<string>
  readonly tags?: ReadonlyArray<Tag>
  readonly status?: 'available' | 'pending' | 'sold'
}

type PetEncoded = {
  readonly id?: bigint
  readonly parent?: ReadonlyArray<PetEncoded>
  readonly internalId?: string
  readonly name: string
  readonly category?: typeof Category.Encoded
  readonly photoUrls: ReadonlyArray<string>
  readonly tags?: ReadonlyArray<typeof Tag.Encoded>
  readonly status?: 'available' | 'pending' | 'sold'
}

export const Pet: Schema.Codec<Pet, PetEncoded> = Schema.Struct({
  id: Schema.optionalKey(Schema.BigInt.annotate({ format: 'int64', examples: [BigInt('10')] })),
  parent: Schema.optionalKey(Schema.Array(Schema.suspend((): Schema.Codec<Pet, PetEncoded> => Pet))),
  internalId: Schema.optionalKey(Schema.String.check(Schema.isPattern(new RegExp('^[0-9]{1,19}$'))).annotate({ examples: ['10'] })),
  name: Schema.String.annotate({ examples: ['doggie'] }),
  category: Schema.optionalKey(Schema.suspend((): Schema.Codec<Category, typeof Category.Encoded> => Category)),
  photoUrls: Schema.Array(Schema.String),
  tags: Schema.optionalKey(Schema.Array(Tag)),
  status: Schema.optionalKey(Schema.Literals(['available', 'pending', 'sold']).annotate({ description: 'pet status in the store' })),
})